api/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions==1.11.2
certifi==2022.6.15
charset-normalizer==2.1.1
et-xmlfile==1.1.0
ez-setup==0.9
idna==3.3
numpy==1.23.2
openpyxl==3.0.10
pandas==1.4.4
pyodbc==4.0.34
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.40
urllib3==1.26.12
